Fix return type on halide_type_t::bytes
Former-commit-id: c8fbe902e7124c724e4258c6c99dd61699463e37

+++ src/runtime/HalideBuffer.h
@@ -621,7 +621,7 @@ public:
         for (int i = 0; i < dimensions(); i++) {
             size /= dim(i).extent();
         }
-        assert(size == ty.bytes() && "Error: Overflow computing total size of buffer.");
+        assert(size == (size_t)ty.bytes() && "Error: Overflow computing total size of buffer.");
     }
 
     /** Allocate memory for this Buffer. Drops the reference to any
